+++ sdk/storage/Azure.Storage.Files.Shares/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 12.9.0-beta.1 (Unreleased)
 - Added support for service version 2020-12-06.
+- Added support for SDK-calculated transactional hash checksums on data transfer.
 
 ## 12.8.0 (2021-09-08)
 - Includes all features from 12.8.0-beta.1 and 12.8.0-beta.2.